IMPALA-8849: fix IllegalStateException with VARCHAR
The bug is that the serialized size wasn't populated
for VARCHAR in a case when it should have been.
It appears a condition was simply not updated when
VARCHAR was added.

Other code assumed that the serialized size was
populated when the other size field was populated,
which is a reasonable invariant. I documented the
invariant in the class and added validation that
the invariant held.

Defining and checking invariants led to discovering
various other minor issues where the sizes were
set incorrect for fixed-length types or not set for
variable-length types:
* CHAR was not consistently treated as a fixed-length type.
* avgSerializedSize_ was not always updated with avgSize_

Testing:
Added a regression test for this specific case. Adding
the assertions resulted in other cases showing up
related bugs.

Change-Id: Ie45e386cb09e31f4b7cdc82b7734dbecb4464534
Reviewed-on: http://gerrit.cloudera.org:8080/14062
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
(cherry picked from commit c3a67b67faaebf18735fa36d35c80d8c11043f7f)

Change-Id: I58cdd5e49bbd7ca49f55901caba83d552dede5b0
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/54537
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>