IMPALA-8851 : Do not throw authorization exception in drop if exists queries
In case a user issues drop query with "if exists" clause and if the object does not exists
currently Impala throws a authorization exception saying that the user does not
persmissions to issue a drop on that non-existent object. This happens because when the
object is dropped all the associated privileges for that object are also dropped. However
this is inconsistent. Even a user which has permissions to list the tables or databases
will get this authorization exception.

This patch adds a change in this behavior so that users do not get authorization exception
if they have enough privileges to (aka list objects of the given type). A user who does
not have privileges still will receive the authorization exception as expected.

Testing:
1. Added a new test case in the sentry tests which confirms that the authorization
exception is not thrown when a drop if exists query is issued on a object which does not
exist

Change-Id: Iba068935e5da92d71e16e2321afdb8e7b781086a