Fix THttpServer to not call the cookie function with an empty cookie
This patch checks if the value passed in the 'Cookie' header to the
http hs2 server is blank, and if so it ignores it.

The reason to do this is so that a client sending an empty cookie
header isn't counted as a failed cookie attempt, which is incorrect.

Change-Id: I04e96fe97baae474a82fd30f2cd55ccce80570b4
Reviewed-on: http://gerrit.cloudera.org:8080/14149
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 3209055ba2b5a05bedf7c9ecf0edd0e06c2525c9)

Change-Id: I6e865d71c342638753ecc96cb45cae7b1d62b1b4
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/55721
Reviewed-by: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Quasar-L0: quasar-precommit User
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>