IMPALA-8917: Remove hostname from webui links if Knox isn't being used
IMPALA-8897 added the hostname to all links on the debug webui in
order to facilitate proxying connections through Apache Knox. This
makes the webui difficult to use in situations where the hostname is
not DNS-resolvable.

This patch fixes this by only including the hostname with links if
Knox proxying is actually being used, which we determine by looking
for the 'x-forwarded-context' header in the request, which Knox adds
to all requests.

It also removes the hidden form fields that were added to support Knox
integration when not being accessed through Knox.

It also adds a class comment on Webserver explaining the requirements
for keeping the webui compatible with Knox.

Testing:
- Added a test that checks that links on the webui are made absolute
  when the 'x-forwarded-context' header is in the request.

Change-Id: Ifcf77058dc6ce1d72422a9e3ca7868cdffacff76
Reviewed-on: http://gerrit.cloudera.org:8080/14199
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit de77c61f383794601c79d9783d36235482740417)

Change-Id: I283cc3cbb6a54ea60296890668bd048fc836b7e3
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/56708
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>