CDPD-3137: Add build-kudu-for-impala.sh script
Impala needs to build against Kudu that has been built with
the native toolchain. In the Cauldron build, this was done
by a build-kudu-for-impala.sh script in the cdh repo. The CDP
build uses a different build system that does not use the cdh
repo. So, this takes the original build-kudu-for-impala.sh
script, removes Cauldron-specific steps, and adds it to the
Impala repository.

The differences from the cdh version are:
 - Have the caller pass in a Kudu source directory via the KUDU_HOME
   environment variable rather than using a Cauldron specific path
 - Do not incorporate the GBN into the tarball name

This is something that could go upstream if we found it useful.

Testing:
 - Ran it locally on my machine
 - Ran a Canary build that calls this script and uses the resulting
   Kudu to build Impala

Change-Id: I73f854dde4af32f425155e44334d2c0d1429fd50
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/56734
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: Joe McDonnell <joemcdonnell@cloudera.com>