IMPALA-8946: Fix histogram rendering to Prometheus
Prometheus has two formats that allows to calculate quantiles (on client
side): Histograms and Summaries. The former requires to pick a set of
buckets, while the latter only requires to specifiy which quantiles
(percentiles) the application wants to provide. The difference between
summaries and histograms on Prometheus can be found here:
https://prometheus.io/docs/practices/histograms/

Previous to this changes, Impala histograms were rendered as Prometheus
histograms, but it was calculating percentiles instead of bucketing the
recorded values, which is not what Prometheus clients expects.

From now on, Impala histograms will be rendered as Prometheus summaries
instead, which expects quantiles.

To be compliant with Prometheus summaries, a `_sum` serie needs to
be added, containing the sum of all the registered values.

The current `HdrHistogram` implementation on be::utils does not have
the `TotalSum` method (and the required properties to track it). I
copied the implementation from be::kudu::utils and added a couple of
unit tests which covers the sum method.

Also adapted the Prometheus tests to verify that sum serie is properly
rendered.

Change-Id: I8849586964140c040ca4ea14b777c2d98d126b59
Reviewed-on: http://gerrit.cloudera.org:8080/14280
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
(cherry picked from commit f08d21335c5a81d31a3de7dc838ac4487080ac42)

Change-Id: Idb9e0d9374f650ccb9e961ea7c3af8f3e5dac9ee
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/58682
Reviewed-by: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>