[cloudera-build] Add cloudera/cdp_install_cmd.sh script
This moves the logic from Impala's install_cmd section
in CDPD's components.ini into cloudera/cdp_install_cmd.sh.
There is a corresponding change in CDPD to call this
script and preserve the corresponding artifacts.

The script is not a complete translation of the CDPD
code:
 - It creates a Kudu jars tarball (needed for upstream)
 - It adds code to dump information about ccache

As a cleanup, this moves build-kudu-for-impala.sh to
the cloudera directory.

Change-Id: If7d77b3139f8707634774f812c59a030be1f2d5c
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/58146
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>
Quasar-L0: Joe McDonnell <joemcdonnell@cloudera.com>