IMPALA-2138: part 3: refactor analytic planning
This is a minor refactor of the analytic planner that moves
the logic to generate a "buffered tuple descriptor" and
equality expressions to the AnalyticEvalNode.

This is required for projection, because the buffered tuple
layout is required to match the input tuple layout, so needs
to be updated if the input tuple is projected.

After the refactoring, this can be correctly updated simply
by regenerating the tuple descriptor and expressions,
instead of attempting to "fix up" the existing expressions
with smaps, etc.

I think the plan node is actually a more approprate place for
this logic regardless, since it's tightly entwined with the
backend's assumption that the two tuple layouts are identical.

Change-Id: I691f1509a85cad51d2654278c0a4ba70a10b2706
Reviewed-on: http://gerrit.cloudera.org:8080/14325
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit 7b3da1d6d27320be92839d4594896b7515f2f91f)

Change-Id: Ib4e8c3aa5cb9137a23cbdedcf39fea37b0028961
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/58953
Reviewed-by: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Quasar-L0: quasar-precommit User
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>