IMPALA-9082: make WebserverTest error checking stricter
WebserverTest::TestWithSpnego has been flaky lately, but I have been
unable to repro it. This patch is an attempt to make it easier to
debug the issue the next time it shows up in automated builds.

The test performs a GET that is expected to fail and then checks that
the metrics show the failed GET. The flaky failures occur when the
metrics do not show a failed attempt.

It appears that what's happening is the GET is failing before actually
reaching the webserver. However, because the GET is expected to fail
and because we only verify that it did fail by checking that HttpGet()
returned some error status, whatever unexpected error is occuring is
getting lost.

This patch instead checks the actual text of the error that is
returned by HttpGet() to make sure it is correct and logs the error if
it isn't.

Change-Id: I820336271cf25130538ceae2eed10a72a73d2adc
Reviewed-on: http://gerrit.cloudera.org:8080/14672
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
(cherry picked from commit a4743fd9bb9158ac5c4de55f77276f39b4570346)

Change-Id: I20faff2c4104afe3995c0042e602f1f1ad946294
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62334
Reviewed-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>