Fix impala-config-branch.sh behavior when CDP_GBN is set explicitly
When someone sets CDP_GBN explicitly, impala-config-branch.sh gets
the cdh-root.properties file from the GBN's s3 location. This
was failing, because it was not creating the directory before
downloading the file to that location. This works normally because
other parts of the file create that directory.

This creates the directory before running wget.

Change-Id: Ia25d3d26b68249b87f3d2dac706fb2f27b0356cd
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/60712
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Quasar-L0: quasar-precommit User
Reviewed-by: Anurag Sai Raghav Krishna Mantripragada <anurag@cloudera.com>