IMPALA-9102: Have webserver fall back to --hostname
When the webserver determines its hostname, it first uses the value of
--webserver_interface. If this is a wildcard address, then it uses the
value returned by gethostname().

In some setups, gethostname() may return the short name and not the
fqdn. This is likely undesirable, eg. because this value is used to
qualify links when Knox proxying is being used, in which case the fqdn
is required.

In other places in Impala with the same basic issue, we allow users to
set the desired hostname with the --hostname flag. This patch has the
webserver fall back to the --hostname flag if --webserver_interface is
not set. If neither is set, we'll fall back to gethostname().

Testing:
- Ran existing webserver tests.
- Manually tested in an environment where gethostname() does not
  return the fqdn

Change-Id: I76d9b8b5286aa7cfc72b0ccd43de8dafda7ff9fe
Reviewed-on: http://gerrit.cloudera.org:8080/14568
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
(cherry picked from commit 5dc5035a794c8a5fbd027b2309e598923f8c60f4)

Change-Id: If6a259cc2cca73f0f80febe5e1061c3f3b159a3f
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/61212
Reviewed-by: Jenkins User <jenkins@cloudera.com>
Tested-by: Jenkins User <jenkins@cloudera.com>
CDH-Build: zuul <zuul@cloudera.com>
Quasar-L0: quasar-precommit User
CDH-Build: Jenkins User <jenkins@cloudera.com>
Quasar-L0: Jenkins User <jenkins@cloudera.com>