[cloudera-build] Reduce build concurrency for CDP build
Some runs of the CDP build are seeing g++ getting killed, presumably because
the system runs out of memory. The CDP container runs in a container with
a large number of CPU threads compared to the amount of memory. C++ compilation
for Impala can average around 600+MB for most files, with some files in the
range of 1-2GB.

This reduces concurrency for Impala's CDP build to 75% of the cores on the
box by setting IMPALA_BUILD_THREADS.

Change-Id: I579a2232a12f691fee56d856338e9ef3a5043bc2
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62161
CDH-Build: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>