[cloudera-build] CDPD-5826: Set CCACHE_BASEDIR to improve hit rates
The CDP build relies on reusing a ccache to improve performance.
This is true even across different Jenkins jobs, which means that
the workspace directory may be different. For example:
/grid/0/jenkins/workspace/Job_A/path/to/file
vs
/grid/0/jenkins/workspace/Job_B/path/to/file
Since ccache will hash the full paths, the two will hash to
different values even if they are the same file. To correct
this, the CCACHE_BASEDIR tells ccache to ignore the base part
of the path, so it can get hits even when the jobs use
different directories.

Change-Id: Ia56ca049ab413fb59cd74862051b8b39f113ea22
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62171
CDH-Build: Joe McDonnell <joemcdonnell@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: Bhavik Bavishi <bbavishi@cloudera.com>
Reviewed-by: Bhavik Bavishi <bbavishi@cloudera.com>