[cloudera-build] CDPD-5826 (part 3): Fix CCACHE_BASEDIR for kudu_for_impala build
The previous fixes for CDPD-5826 added ccache settings to allow the cache
to be reused across Jenkins jobs (which use different base directories).
The previous fixes are effective for Impala compilation, but they are not
effective for compiling Kudu with Impala's toolchain.

The problem is that ccache hashes the compiler path as part of the command.
In this case, CCACHE_BASEDIR is set to KUDU_HOME, but the compiler is inside
IMPALA_HOME. This means that the absolute compiler path is not fixed by
CCACHE_BASEDIR. This changes CCACHE_BASEDIR to SOURCE_ROOT for compilation
of Kudu with Impala's toolchain. SOURCE_ROOT is a directory above KUDU_HOME
and IMPALA_HOME, and it incorporates the difference in base directory due
to the Jenkins job name. This uses the same settings for Impala compilation
for simplicity.

Change-Id: I3b3c32587fc554cdee810f96d328e49b8f37166b
Reviewed-on: https://gerrit.sjc.cloudera.com/c/cdh/impala/+/62604
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
CDH-Build: Joe McDonnell <joemcdonnell@cloudera.com>
Quasar-L0: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>